Рефакторинг тестов + исправлены ошибки в файловом хранилище.
Former-commit-id: 5bf22856156829e9a70c323da8ec454ccbb951d7

+++ Bwl.Storage.UniversalORM.Files/FileStorage.vb
@@ -99,9 +99,13 @@ Public Class FileObjStorage
 	Private Function CreateIndex(obj As ObjBase) As Boolean
 		For Each indexing In _indexingMembers
 			Try
-				Dim indexValue = ReflectionTools.GetMemberValue(indexing.Name, obj).ToString
+				Dim indexValue = ReflectionTools.GetMemberValue(indexing.Name, obj)
+				If indexing.Type = GetType(DateTime) Then
+					indexValue = CType(indexValue, DateTime).Ticks
+				End If
+
 				Dim path = GetIndexFileName(obj.GetType, indexing.Name)
-				Dim value = obj.ID + " " + indexValue + vbCrLf
+				Dim value = obj.ID + " " + indexValue.ToString + vbCrLf
 
 				Dim existsID As Boolean = False
 				If File.Exists(path) Then
@@ -170,195 +174,190 @@ Public Class FileObjStorage
 		Return _folder + Utils.Sep + type.Name + "." + index + ".index"
 	End Function
 
-	Private Function SortDictionary(dictionary As Dictionary(Of String, String), sortParam As SortParam) As List(Of String)
+	Private Shared Function SortDictionary(dictionary As Dictionary(Of String, Object), sortParam As SortParam) As IEnumerable(Of String)
 		Dim result As New List(Of String)
-		Dim sortedDictionary As IEnumerable(Of KeyValuePair(Of String, String))
-		If (sortParam.SortMode = SortMode.Ascending) Then
-			sortedDictionary = dictionary.OrderBy(Function(v) v.Value)
-		Else
-			Dim sorted = From pair In dictionary Order By pair.Value Descending
-			sortedDictionary = dictionary.OrderByDescending(Function(v) v.Value)
-		End If
-		For Each item In sortedDictionary
-			result.Add(item.Key)
-		Next
-		Return result
-	End Function
-
-	Private Function Sort(sortParam As SortParam) As List(Of String)
-		Dim result As New List(Of String)
-		Dim dictionary As New Dictionary(Of String, String)
-		Dim indexInfo = _indexingMembers.Find(Function(x) x.Name = sortParam.Field)
-		If indexInfo IsNot Nothing Then
-			Dim indexFileName = GetIndexFileName(Type.GetType(SupportedType.AssemblyQualifiedName), indexInfo.Name)
-			Dim fileReader = My.Computer.FileSystem.OpenTextFileReader(indexFileName)
-			Dim stringReader = String.Empty
-			While fileReader.Peek <> -1
-				stringReader = fileReader.ReadLine()
-				If stringReader <> String.Empty Then
-					Dim line = stringReader.Split(" "c)
-					Dim value = String.Empty
-					For i = 1 To line.Count - 1
-						value += line(i)
-					Next
-
-					dictionary.Add(line(0), value)
-				End If
-			End While
-			fileReader.Close()
-			result = SortDictionary(dictionary, sortParam)
+		If dictionary IsNot Nothing AndAlso dictionary.Any Then
+			If (sortParam.SortMode = SortMode.Ascending) Then
+				result.AddRange(dictionary.OrderBy(Function(pair) pair.Value).Select(Function(pair) pair.Key))
+			Else
+				result.AddRange(dictionary.OrderByDescending(Function(pair) pair.Value).Select(Function(pair) pair.Key))
+			End If
 		End If
 		Return result
 	End Function
 
 	Private Function Sort(list As List(Of String), sortParam As SortParam) As List(Of String)
 		Dim result As New List(Of String)
-		Dim dictionary As New Dictionary(Of String, String)
-		Dim indexInfo = _indexingMembers.Find(Function(x) x.Name = sortParam.Field)
-		If indexInfo IsNot Nothing Then
-			Dim indexFileName = GetIndexFileName(Type.GetType(SupportedType.AssemblyQualifiedName), indexInfo.Name)
-			Dim fileReader = My.Computer.FileSystem.OpenTextFileReader(indexFileName)
-			Dim stringReader = String.Empty
-			While fileReader.Peek <> -1
-				stringReader = fileReader.ReadLine()
-				If stringReader <> String.Empty Then
-					Dim line = stringReader.Split(" "c)
-					Dim value = String.Empty
-					For i = 1 To line.Count - 1
-						value += line(i)
-					Next
-					For Each item In list
-						If item.Contains(line(0)) Then
-							dictionary.Add(line(0), value)
+		If list IsNot Nothing AndAlso list.Any Then
+			Dim dictionary As New Dictionary(Of String, Object)
+			Dim indexInfo = _indexingMembers.Find(Function(x) x.Name = sortParam.Field)
+			If indexInfo IsNot Nothing Then
+				Dim indexFileName = GetIndexFileName(Type.GetType(SupportedType.AssemblyQualifiedName), indexInfo.Name)
+				Dim fileReader = My.Computer.FileSystem.OpenTextFileReader(indexFileName)
+				Try
+					Dim stringReader = String.Empty
+					While fileReader.Peek <> -1
+						stringReader = fileReader.ReadLine()
+						If stringReader <> String.Empty Then
+							Dim line = stringReader
+							Dim startSpacePos = line.IndexOf(" "c)
+							Dim idStr = line.Substring(0, startSpacePos)
+							If list.Contains(idStr) Then
+								Dim valueStr = ""
+								If startSpacePos > 0 Then
+									valueStr = line.Substring(startSpacePos + 1, line.Length - startSpacePos - 1)
+								End If
+								Try
+									Dim value As Object = Nothing
+									If indexInfo.Type = GetType(DateTime) Then
+										value = New DateTime(Convert.ToInt64(valueStr))
+									Else
+										value = CTypeDynamic(valueStr, indexInfo.Type)
+									End If
+									dictionary.Add(idStr, value)
+								Catch ex As Exception
+								End Try
+							End If
 						End If
-					Next
-				End If
-			End While
-			fileReader.Close()
-			result = SortDictionary(dictionary, sortParam)
+					End While
+				Finally
+					fileReader.Dispose()
+				End Try
+				result.AddRange(SortDictionary(dictionary, sortParam))
+			End If
 		End If
 		Return result
 	End Function
 
 	Public Overrides Function FindObj(searchParams As SearchParams) As String()
-		If searchParams Is Nothing Then Return FindAllObjs()
-
-		Dim result As New List(Of String)()
-		Dim tmpResult As New List(Of String)()
-		Dim listResults As New List(Of List(Of String))()
-
-		If searchParams.FindCriterias IsNot Nothing Then
+		Dim result As New List(Of String)
+		If searchParams Is Nothing OrElse searchParams.FindCriterias Is Nothing Then
+			result.AddRange(FindAllObjs())
+		Else
+			Dim listResults As New List(Of List(Of String))()
 			For Each crit In searchParams.FindCriterias
-				tmpResult.Clear()
+				Dim tmpResult As New List(Of String)()
 				Dim indexFileName = String.Empty
 				If searchParams.FindCriterias IsNot Nothing Then
 					Dim indexInfo = _indexingMembers.Find(Function(x) x.Name = crit.Field)
 					If indexInfo IsNot Nothing Then
-
 						indexFileName = GetIndexFileName(Type.GetType(SupportedType.AssemblyQualifiedName), indexInfo.Name)
 						Dim fileReader = My.Computer.FileSystem.OpenTextFileReader(indexFileName)
-						Dim stringReader = String.Empty
-						While fileReader.Peek <> -1
-							stringReader = fileReader.ReadLine()
-							If stringReader <> String.Empty Then
-								Dim line = stringReader.Split(" "c)
-								Dim value = String.Empty
-								For i = 1 To line.Count - 1
-									value += line(i)
-								Next
-								Select Case (crit.Condition)
-									Case FindCondition.eqaul
-										If (value = crit.Value) Then tmpResult.Add(line(0))
-									Case FindCondition.greater
-										If (value > crit.Value) Then tmpResult.Add(line(0))
-									Case FindCondition.greaterOrEqual
-										If (value >= crit.Value) Then tmpResult.Add(line(0))
-									Case FindCondition.less
-										If (value < crit.Value) Then tmpResult.Add(line(0))
-									Case FindCondition.lessOrEqual
-										If (value <= crit.Value) Then tmpResult.Add(line(0))
-									Case FindCondition.likeEqaul
-										Throw New NotSupportedException
-									Case FindCondition.notEqual
-										If (value <> crit.Value) Then tmpResult.Add(line(0))
-								End Select
-							End If
-						End While
-						fileReader.Close()
-						listResults.Add(tmpResult.Select(Function(x) x.Clone().ToString()).ToList())
+						Try
+							Dim stringReader = String.Empty
+							While fileReader.Peek <> -1
+								stringReader = fileReader.ReadLine()
+								If stringReader <> String.Empty Then
+									Dim line = stringReader
+									Dim startSpacePos = line.IndexOf(" "c)
+									Dim idStr = line.Substring(0, startSpacePos)
+									Dim valueStr = ""
+									If startSpacePos > 0 Then
+										valueStr = line.Substring(startSpacePos + 1, line.Length - startSpacePos - 1)
+									End If
+
+									Try
+										Dim value As Object = Nothing
+										If indexInfo.Type = GetType(DateTime) Then
+											value = New DateTime(Convert.ToInt64(valueStr))
+										Else
+											value = CTypeDynamic(valueStr, indexInfo.Type)
+										End If
+
+										Select Case (crit.Condition)
+											Case FindCondition.eqaul
+												If (value = crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+											Case FindCondition.greater
+												If (value > crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+											Case FindCondition.greaterOrEqual
+												If (value >= crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+											Case FindCondition.less
+												If (value < crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+											Case FindCondition.lessOrEqual
+												If (value <= crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+											Case FindCondition.likeEqaul
+												Throw New NotSupportedException
+											Case FindCondition.notEqual
+												If (value <> crit.Value) Then
+													tmpResult.Add(idStr)
+												End If
+										End Select
+									Catch ex As Exception
+									End Try
+								End If
+							End While
+						Finally
+							fileReader.Dispose()
+						End Try
+						listResults.Add(tmpResult)
 					Else
 						MessageBox.Show(String.Format("Указанный тип ({0}) не является индексируемым"), searchParams.SortParam.Field)
 					End If
 				End If
 			Next
-		End If
-		If listResults.Count = 1 Then
-			result = listResults(0)
-		ElseIf listResults.Count > 1 Then
-			'находим список с минимальной длиной для сокращения перебора
-			Dim minres As List(Of String) = Nothing
-			For Each item In listResults
-				If (minres Is Nothing Or ((minres IsNot Nothing) AndAlso (minres.Count > item.Count))) Then
-					minres = item
-				End If
-			Next
-			'сравнение результатов поиска по каждому критерию
-			For Each id In minres
-				Dim exists As Boolean = True
-				For i = 0 To listResults.Count - 1
-					If ((listResults IsNot minres) AndAlso Not (listResults(i).Contains(id))) Then
-						exists = False
-						Exit For
+			If listResults.Count > 0 Then
+				'сравнение результатов поиска по каждому критерию
+				For Each id In listResults(0)
+					Dim exists = True
+					For i = 1 To listResults.Count - 1
+						If Not listResults(i).Contains(id) Then
+							exists = False
+							Exit For
+						End If
+					Next
+					If (exists) AndAlso Not result.Contains(id) Then
+						result.Add(id)
 					End If
 				Next
-				If (exists) AndAlso Not result.Contains(id) Then
-					result.Add(id)
-				End If
-			Next
-		End If
-
-		If searchParams.SortParam IsNot Nothing Then
-			If result.Count > 0 Then
-				result = Sort(result, searchParams.SortParam)
-			Else
-				result = Sort(searchParams.SortParam)
 			End If
 		End If
 
-		If searchParams.SelectOptions IsNot Nothing Then
+		If searchParams IsNot Nothing AndAlso searchParams.SortParam IsNot Nothing Then
+			If result.Count > 0 Then
+				result = Sort(result, searchParams.SortParam)
+			End If
+		End If
+
+		If searchParams IsNot Nothing AndAlso searchParams.SelectOptions IsNot Nothing Then
 			If searchParams.SelectOptions.SelectMode = SelectMode.Top Then
 				'Top
 				If result.Count > 0 Then
-					result = result.GetRange(0, Convert.ToInt32(searchParams.SelectOptions.TopValue))
-				Else
-					result = FindAllObjs().ToList().GetRange(0, Convert.ToInt32(searchParams.SelectOptions.TopValue))
+					Dim topVal = Convert.ToInt32(searchParams.SelectOptions.TopValue)
+					result = result.GetRange(0, topVal)
 				End If
 			Else
 				'Between
 				If result.Count > 0 Then
-					result = result.GetRange(Convert.ToInt32(searchParams.SelectOptions.StartValue),
-															 Convert.ToInt32(searchParams.SelectOptions.EndValue - Convert.ToInt32(searchParams.SelectOptions.StartValue) + 1))
-				Else
-					result = FindAllObjs().ToList().GetRange(Convert.ToInt32(searchParams.SelectOptions.StartValue),
-															 Convert.ToInt32(searchParams.SelectOptions.EndValue - Convert.ToInt32(searchParams.SelectOptions.StartValue) + 1))
+					Dim startVal = Convert.ToInt32(searchParams.SelectOptions.StartValue)
+					Dim endVal = Convert.ToInt32(searchParams.SelectOptions.EndValue)
+					result = result.GetRange(startVal, endVal - startVal + 1)
 				End If
 			End If
 		End If
+
 		Return result.ToArray()
 	End Function
 
 	Public Overrides Function FindObjCount(searchParams As SearchParams) As Long
 		Dim res As Long = 0
-		If Utils.TestFolderFsm(_folder) Then
-			Dim files = IO.Directory.GetFiles(_folder, "*.obj.json")
-			If files IsNot Nothing Then
-				res = files.Length
-			End If
+		Dim tmp = FindObj(searchParams)
+		If tmp IsNot Nothing AndAlso tmp.Any Then
+			res = tmp.Length
 		End If
 		Return res
 	End Function
 
-	Private Function FindAllObjs() As String()
+	Private Function FindAllObjs() As IEnumerable(Of String)
 		Dim result As New List(Of String)
 		If Utils.TestFolderFsm(_folder) Then
 			Dim files = IO.Directory.GetFiles(_folder, "*.obj.json")
@@ -367,7 +366,7 @@ Public Class FileObjStorage
 				result.Add(fileParts(fileParts.Length - 3))
 			Next
 		End If
-		Return result.ToArray
+		Return result
 	End Function
 
 	Public Overrides Function GetObj(id As String) As ObjBase
@@ -375,13 +374,11 @@ Public Class FileObjStorage
 		If Utils.TestFolderFsm(_folder) Then
 			Dim file = GetFileName(id)
 			Try
-
 				If IO.File.Exists(file) Then
 					Dim str = IO.File.ReadAllText(file, Utils.Enc)
 					Dim oi = CType(CfJsonConverter.Deserialize(str, GetType(ObjInfo)), ObjInfo)
 					obj = CType(CfJsonConverter.Deserialize(oi.Obj, oi.ObjType), ObjBase)
 				End If
-
 			Catch ex As Exception
 				Dim err = "Err FileStorage.GetObj _ file: " + file + vbCrLf + ex.ToString
 				Dim ex1 = New InvalidOperationException(err, ex)
@@ -424,7 +421,15 @@ Public Class FileObjStorage
 	End Sub
 
 	Public Overrides Sub RemoveAllObjects()
-		Directory.Delete(_folder, True)
-		Utils.TestFolderFsm(_folder)
+		Try
+			If Directory.Exists(_folder) Then
+				Directory.Delete(_folder, True)
+			End If
+		Catch ex As Exception
+		End Try
+		Try
+			Utils.TestFolderFsm(_folder)
+		Catch ex As Exception
+		End Try
 	End Sub
 End Class
